requirements.txt
Django==1.5.1
psycopg2==2.5.1
South==0.7.6
misaka==1.0.2
django-model-utils==1.4.0
